Move writing-related files to sfta/writing/
